Homotopy Continuation For Tensor Decomposition

This is a short and rudimentary user guide for Homotopy Continuation For Tensor Decomposition. This package comes with my master's thesis and detailed motivations for the algorithms available in this package can be found in the thesis text.

Getting Started

Installing the package

Since this is not an official package, it has to be installed manually.

First, load all the dependencies via Julia's package manager.

using Pkg

Pkg.activate("path/to/HomotopyContinuationForTensorDecomposition")

Pkg.instantiate()

Secondly, we should install HomotopyContinuationForTensorDecomposition. There are two ways to do this.

Add HomotopyContinuationForTensorDecomposition.jl to Julia's LOAD_PATH.

push!(LOAD_PATH,"path/to/src/HomotopyContinuationForTensorDecomposition.jl")

using HomotopyContinuationForTensorDecomposition

Use include().

include("path/to/src/HomotopyContinuationForTensorDecomposition.jl")

using .HomotopyContinuationForTensorDecomposition

A first example

The example below shows how to interact with the implemented algorithms.

using LinearAlgebra: norm

dims = (43, 8, 7); # dimensions of the tensor
r = 5; # tensor rank

X, Y, Z, T = cpd_gen(dims, r); # generate a random tensor

Xcalc, Ycalc, Zcalc = cpd(T, dims, r);

Tcalc = cpd_gen(Xcalc, Ycalc, Zcalc);

norm(Tcalc - T)/norm(T) # backward relative error

3.0987487849154017e-15

The implemented algorithms only work for third order tensors with dimensions (n₁, n₂, n₃) for which it holds that r < min(n₁, n₂n₃) and n₁ ≧ n₂ ≧ n₃. The implemented methods are:

kl_cpd (cpd via the method described by Kuo, Lee [1]);

kld_fd_cpd (cpd via the dual problem by reducing the size of the square system);

kld_pr_cpd (cpd via the dual problem by a parameter homotopy);

kl_lr_cpd (Kuo and Lee's method optimised for low ranks);

cpd (wrapper which selects the optimal method from those above).

[1] Y.-C. Kuo and T.-L. Lee. Computing the unique candecomp/parafac decomposition of unbalanced tensors by homotopie method. Linear Algebra and its applications, 556:238–264, 2018.

Benchmarking

The file benchmarks/Benchmark.jl contains a small module to test the performance of the different algorithms. Additionally, an interface with Tensorlab2016[2] is provided via the MATLAB package. The following four steps explain how to run some tests.

Set Julia's working directory to ../benchmarks.

cd("path/to/HomotopyContinuationForTensorDecomposition/benchmarks")

Put a copy of Tensorlab2016 in ../benchmarks.

Load the Benchmarks module.

include("Benchmark.jl")

using .Benchmarks

Call the function run_benchmark().

run_benchmark().

This starts a command line interface through which it is possible to provide all the experiment settings. Should the options provided there be insufficient, experiments can be created manually and passed as an argument to run_benchmark.

?Benchmarkinstance # show documentation for the benchmarkinstance

# every experiment is represented by a Benchmarkinstance
bmis = Benchmarkinstance[]

dims = (43, 8, 7);
r = 15;
model = 1;
s = 2.0;
c = 1.0;

push!(bmis, Benchmarkinstance(dims, r; model=model, s=s, c=c));

run_benchmark(bmis); # run the experiments in `bmis`
# ...

The experiment results will be saved in a .csv file in the folder ../data. Before and after the experiment run_benchmark() prints the exact file name. Additionally, the measurements are printed during the experiment.

[2] Vervliet N., Debals O., Sorber L., Van Barel M. and De Lathauwer L. Tensorlab 3.0, Available online, Mar. 2016. URL: https://www.tensorlab.net/

CPD methods

HomotopyContinuationForTensorDecomposition.cpd — Method

cpd(T, dims, r)

Chooses and calls the most efficiënt algorithm from

kl_cpd

kld_fd_cpd

kld_pr_cpd

kl_lr_cpd

by taking into acount the dimensions and the rank.

if r < n₂ and r < n₃: kl_lr_cpd

if r < n₂ or r < n₃: kld_fd_cpd

if r <= n₂+n₃-2: kld_pr_cpd

if r < n₁: kl_cpd

If none of the above conditions are met, cpd throws an assertion error.

KuoLee.kl_cpd — Method

kl_cpd

Calculate the cpd of an order 3 tensor via the method described in Kuo, Lee.

Y.-C. Kuo and T.-L. Lee. Computing the unique candecomp/parafac decomposition of unbalanced tensors by homotopie method. Linear Algebra and its applications, 556:238–264, 2018.

Inputs

T: flattened real or complex tensor, type Array{d_type, 2}

r: rank of the tensor, type Int64

dims: tuple of the dimensions, first dimension should be larger than r, type Tuple{Int64, Int64, Int64}.

Outputs

The method outputs the calculated factor matrices

\[ T \approx (Z_{\text{calc}} \odot Y_{\text{calc}})X_{\text{calc}}^H\]

Xcalc

Ycalc

Zcalc

Example

dims = (43, 8, 7);
r = 17;
X, Y, Z, T = cpd_gen(dims, r);

Xcalc, Ycalc, Zcalc = kl_cpd(T, r, dims);

T̃ = cpd_gen(Xcalc, Ycalc, Zcalc);
ϵ = norm(T-T̃)/norm(T);

KuoLee.kl_lr_cpd — Method

kl_lr_cpd

Calculate the cpd of an order 3 unbalanced flattened tensor via the method described in Kuo, Lee.

Y.-C. Kuo and T.-L. Lee. Computing the unique candecomp/parafac decomposition of unbalanced tensors by homotopie method. Linear Algebra and its applications, 556:238–264, 2018.

This method is optimised towards lower ranks: r < n₂ and r < n₃. The flattened tensor has the following decomposition:

\[ T = (Z \odot Y)X^H = EF^H.\]

As described in the thesis, this method exploits the information that the linear subspaces 𝐘 and 𝐙 are r dimensional and known. Therefor we search a basis Y for 𝐘 and a basis Z for 𝐙 s.t. Z ⊙ Y is a basis for E.

Inputs

T: flattened real or complex tensor, type Array{d_type, 2}

r: rank of the tensor, type Int64

dims: tuple of the dimensions, first dimension should be larger than r, type Tuple{Int64, Int64, Int64}.

Outputs

The method outputs the calculated factor matrices

\[ T \approx (Z_{\text{calc}} \odot Y_{\text{calc}})X_{\text{calc}}^H\]

Xcalc

Ycalc

Zcalc

Dual.kld_fd_cpd — Method

kld_fd_cpd

Calculate the cpd of a flattened order 3 tensor via the dual problem. The algoritm numerically reduces the size of the square system used for homotopy continuation by setting some variables to a random value. This method solves the dual problem by calling dual_solve and than translates the dual problem to the primal via translate_dual_prim. The documentation of these methods elaborates more on the techniques used.

Inputs

T: flattened real or complex tensor

r: rank of the cpd

dims: tuple of the dimensions, first dimension should be larger than r.

Outputs

The method outputs the calculated factor matrices

\[ T \approx (Z_{\text{calc}} \odot Y_{\text{calc}})X_{\text{calc}}^H\]

Xcalc

Ycalc

Zcalc

Optional arguments

maxit_basis: (Int64) maximum number of iterations used in solve_dual.

maxit_ortho: (Int64) maximum number of iterations used in translate_dual_prim.

verbose: (Bool) give some information about the number of iterations / the type of problems solved.

strategy: (String) tell solve_dual which methods it should use to grow the basis. See solve_dual for the possible values.

Example

dims = (43, 8, 7);
r = 6;
X, Y, Z, T = cpd_gen(dims, r);

Xcalc, Ycalc, Zcalc = kld_fd_cpd(T, r, dims);

T̃ = cpd_gen(Xcalc, Ycalc, Zcalc);
ϵ = norm(T-T̃)/norm(T);

Dual.kld_pr_cpd — Method

kld_pr_cpd

Calculate the cpd of an order 3 unbalanced flattened tensor. This method uses a parameter homotopy to solve the dual problem. This function constructs an initial decomposition and then calls kld_pr_cpd_refine.

Inputs

T₁: flattened tensor

r: rank of T₁

dims: tuple with dimensions (n₁, n₂, n₃) of the tensor such that T₁ has dimensions n₂n₃×n₁

Outputs

The method outputs the calculated factor matrices

\[ T \approx (Z_{\text{calc}} \odot Y_{\text{calc}})X_{\text{calc}}^H\]

Xcalc

Ycalc

Zcalc

Dual.kld_pr_cpd_refine — Method

kld_pr_cpd_refine

Calculate the cpd of an unbalanced flattened order 3 tensor via the dual problem. The method solves the dual problem by paramter homotopy and requires an initial decomposition. See kl_pr_cpdfor a wrapper which also constructs the initial decomposition.

Inputs

T₀: start flattened tensor

T₁: target flattened tensor

Y₀: factor matrix of start tensor

Z₀: factor matrix of start tensor

dims: dimensions (n₁, n₂, n₃) of start and target tensor. Y₀ should have dimensions n₂×r, Z₀ should have dimensions n₃×r.

r: rank of the tensor

Outputs

The method outputs the calculated factor matrices

\[ T \approx (Z_{\text{calc}} \odot Y_{\text{calc}})X_{\text{calc}}^H\]

Xcalc

Ycalc

Zcalc

Benchmarks

Benchmarks.run_benchmark — Function

run_benchmark

Script which will run experiments. The script prompts for user feedback after printing the settings to the shell. This allows the user to still throw an interrupt exception using ctrl-C.

The following variables have to be manually inserted in Benchmarks.jl

nb_warmup: number of iterations before we take a measurement.

nb_samples: number of measurements to take for each experiment (or benchmark_instance).

benchmark_set: list of experiments (benchmark_instance) to perform.

kl: (Bool) indicate if we run kl_cpd. This is mostly the most expensive method.

Benchmarks.Benchmarkinstance — Type

Benchmarkinstance

Struct which represents an experiment to perform.

Contains

n₁: first dimension of tensor.

n₂: second dimension of tensor.

n₃: third dimension of tensor.

r: rank of the tensor.

η: parameter to controll noise. The noise is applied to the tensor to approximate the empirical condition number.

model: the model used to construct the tensor. The models are described in [3]

s: parameter used in model 1 and model 2

c: parameter used in model 1

sample_fm: can only be false if the tensor format is perfect. In that case a random tensor is generated directly. If true the tensor is generated form random factor matrices.

[3] Breiding and Vannieuwenhoven, A Riemannian trust region method for the canonical tensor rank approximation problem

Helper functions

DualUtils.dual_solve — Method

dual_solve

Calculates a rank-one basis for Null(E') by intersecting the positive dimensional variety with hyperplanes. The dimensions of E are n₂n₃ × r. Requires that r < n₂ + n₃ - 2.

All solutions (𝐮ⱼ, 𝐯ⱼ) satisfy

\[ (\mathbf{u}_j \odot \mathbf{v}_j)^H\mathbf{e}_i = 0, \quad \forall i=1,\dots, r,\quad \forall j=1,\dots, n_2n_3-r.\]

This method is a wrapper for constructing random systems, solving them and afterwards parsing and checking the solutions. Three different solvers are used:

Linpolysolve(): if r < n₂ or r < n₃ it is possible to contruct linear systems in r variables. If this is done depends on the given strategy.

Full degree solver: calculates all solutions of a nonlinear system using HomotopyContinuation.

dual_paramsolve: calculates a set of solutions using, then tracks these solutions while the parameters of the system (those which define the hyperplanes) are changed.

Using the default strategy 'xlastxy' (old naming convention) the method will start by dehomogenizing the system such that it is linear in 𝐯 if r < n₂. When no new independent solutions are found, it will use r variables in 𝐯 and one in 𝐮. In the case that r > n₂, all variables in 𝐯 and r-n₂ variables in 𝐳 are used.

Inputs

E basis for the space 𝐄 to which we seek orthogonal rank-one vectors.

dims dimensions of the tensor.

r rank of the tensor.

Outputs

success: integer flag which equals 1 if all went well.

N: the rank-one basis for 𝐍 = Nul(𝐄)

UV: container with the vectors 𝐮 and 𝐯 such that 𝐮ⱼ ⊙ 𝐯ⱼ = 𝐧ⱼ. The vector 𝐮ⱼ can be accessed as UV[j][2] and the vector 𝐯ⱼ as UV[j][1].

Keyword arguments

verbose: print some info to the command line

strategy: indicate a strategy to numerically eliminate variables.

Strategies

The strategy can be set via the keyword argument strategy. Possible values are

!!! note The strategy names come from an old naming convention.

"x_last_xy": if r < n₂ eliminate all variables in 𝐮 such that the dual problem is linear in 𝐯. Once this doesn't result in new solutions or r >= n₂, solve non-linear systems with one variable in 𝐯 and r variables in 𝐮 with the full degree solver.

"y_last_yx": if r < n₃ eliminate all variables in 𝐯 such that the dual problem is linear in 𝐮. Once this doesn't result in new solutions or r >= n₃, solve non-linear systems with one variable in 𝐮 and r variables in 𝐯 with the full degree solver.

"optimal": choose automatically between "x_last_xy" and "y_last_yx" such that linear systems are used when r < n₂ or r < n₃.

"x_last_xypar": (not properly tested) similar to "x_last_xy" but dual_paramsolve is used to solve the nonlinear systems.

"y_last_yxpar": (not properly tested) similar to "y_last_yx" but dual_paramsolve is used to solve the nonlinear systems.

Possible improvements:

Optimise memory allocation. Most memory is still allocated 'on the fly'.

Uniform naming conventions in all helper functions.

DualUtils.rankfact_nullspace — Method

rankfact_nullspace

Calculate the full rank factorisation of T s.t. T = EF' and N = null(E').

Inputs

T: matrix to factorize

r: rank of t

Outputs

E: matrix such that T = EF'

F: matrix such that T = EF'

N: matrix such that N'E = 0.

DualUtils.translate_dual_prim — Method

translatedualprim

Calculate a rank-one basis Z ⊙ Y orthogonal to U ⊙ V. The dimensions are

U: n₃ × n₂n₃-r;

V: n₂ × n₂n₃-r;

Z: n₃ × r;

Y: n₂ × r.

This solutions (𝐲ᵢ, 𝐳ᵢ) satisfy

\[ (\mathbf{u}_j \odot \mathbf{v}_j)^H(\mathbf{z}_i \odot \mathbf{y}_i) = 0, \quad \forall i=1,\dots, r,\quad \forall j=1,\dots, n_2n_3-r.\]

Moreover the unless soft is activated the method asserts that the returned solutions are zero-dimensional.

Inputs

V: see above.

U: see above.

dims: see above.

r: see above.

Keyword arguments

verbose: (true/false) provide some output on the number of iterations.

maxit: (int) maximum number of iterations.

sigma_threshold: (float) decides when a new vector is considered to be part N. If σ[end]/σ[1] < sigma_threshold the solution is rejected due to linear dependency.

soft: (true/false) silences assertions for zero-dimensional solution components. Thus, activating soft will not interrupt the software if a positive dimensional solution component is encountered.

Outputs

success: flag which equals the interger one if the method succeeded.

YZ: container for 𝐲ᵢ and 𝐳ᵢ which can be accessed as YZ[i][1] and YZ[i][2].

Possible improvements

reduce the number of calls to nullspace.

DualUtils.combis — Method

combis

Experimental function to generate some combinatorics. Creates different combinations of select_nb elements out of select_list. The method stops if it can find no new samples or if the number of combinations reaches nb_samples.

DualUtils.dual_constructvars — Method

dual_constructvars

Helper function for the dual_solve() method. The function constructs two arrays of variables 𝐯 and 𝐮 which have a random set of elements set to a random value.

!!! note Internally this method uses a different naming convention.

Inputs

n₂: (int) dimension of 𝐯 (corresponding tensor has dimensions n₁×n₂×n₃).

n₃: (int) dimension of 𝐮.

r: (int) rank of the corresponding tensor.

verbose: print some information about the proceedings to stdout.

setuv: (string) defines which variables should be set.

X_combos: (old naming convention, V_combos would be more appropriate) list of different combinations of r numbers from the range 1:n₂.

Y_combos: (old naming convention, U_combos would be more appropriate) list of different combinations of r numbers from the range 1:n₃:

Xcomb_index: (old naming convention) tells the method which 'combo' of X_combos should be used.

Ycomb_index: (old naming convention) similar as above.

it: global iteration counter. Used to only print a message during the first iteration if verbose equals true.

Outputs

x: (old naming convention, 𝐯 would be more appropriate) array consisting of PolyVar types and / or random values.

y: (old naming convention, 𝐮 would be more appropriate) similar as above.

t: if the polynomial system is solved with dual_paramsolve, then t represents the PolyVars in x and y which would otherwise be set to a random value.

add_eqs: if there are more than r variables (PolyVar), then add_eqs communicates this further.

rm_eqs: depreciated

setuv: (string) communicates which variables are set.

vars: array of all variables in the polynomial system (only 'true' variables, so no parameters such as t).

DualUtils.dual_deparsevars — Method

dual_deparsevars

Put the solutions 𝛇 of the polynomial system at the right place in 𝐮 and 𝐯. Can be interpreted as the inverse function of dual_constructvars.

!!! note Still uses an old naming convention internally.

DualUtils.dual_filtsols — Method

# dual_filtsols

Filter all solutions in 𝐧_list such that

their residual is low enough

they are not trivial

if added to the basis N, the vectors are independent enough.

DualUtils.dual_paramsolve — Method

dual_paramsolve

The polynomial system is parametrised in t and u. This method first calculates a solution set for a random t0 and u0 after which it tracks those solutions to random t1 and u1, then it tracks those to t2 and u2 and so on.

DualUtils.dual_solvepolysys — Method

dual_solvepolysys

Wrapper which based on given settings selects the right solver. It can call four different methods

dual_paramsolve: track a set of solutions to different locations in parameter space.

HomotopyContinuation.solve(): full degree homotopy.

HomotopyContinuation.solve(…, variable_groups= …): full degree homotopy with m-homogeneous system.

linpolysolve: solves a linear polynomial system.

DualUtils.dual_track_single — Method

# dual_track_single

Track all solutions of the dual start problem to the target problem:

start condition: U₀ ⊙ V₀ ⟂ E₀,

target condition: U₁ ⊙ V₁ ⟂ E₁.

Inputs

UV₀: container for the start solutions.

E₀: as explained above.

E₁: as explained above.

dims: (n₁, n₂, n₃) dimensions of the tensor. E₀ and E₁ have dimension n₂n₃×r.

r: rank of the tensor.

Outputs

success: flag

UV₁: container for target solutions

DualUtils.linpolysolve — Method

linpolysolve

Solves a system of degree one polynomials.

KuoLeeUtils.constrpoly — Method

constrpoly

Construct the polynomial system for the algorithm kl_cpd.

Inputs

N: space to which orthogonal rank-one vectors are sought

dims: tensor dimensions

vars: polynomial variables

Outputs

𝐩: array of polynomial equations

KuoLeeUtils.filtsols — Method

filtsols

Filter solutions which are orthogonal to every vector in N.

Inputs

𝛇: array of solution vectors

N: basis of space to which orthogonal vectors are sought

dims: dimensions of the tensor

r: rank of the tensor

Outputs

Y_filt: array of vectors 𝐲ᵢ (1≦i≦r) of length r such that residual of (𝐳ᵢ ⊙ 𝐲ᵢ)ᴴ𝐧ⱼ is smallest.

Z_filt: array of vectors 𝐳ᵢ (1≦i≦r) of length r such that residual of (𝐳ᵢ ⊙ 𝐲ᵢ)ᴴ𝐧ⱼ is smallest.

KuoLeeUtils.normbalance — Method

Normbalance

Distribute the norm equally over all factor vectors. Put all angle in X.

Inputs

X

Y

Z

Outputs

X: factor matrix X s.t. ||𝐱ᵢ|| = ||𝐲ᵢ|| = ||𝐳ᵢ|| ∠xᵢ₁ = 0. ∀ i,

Y: factor matrix Y s.t. ||𝐱ᵢ|| = ||𝐲ᵢ|| = ||𝐳ᵢ|| ∠yᵢ₁ = 0. ∀ i,

Z: factor matrix Z s.t. ||𝐱ᵢ|| = ||𝐲ᵢ|| = ||𝐳ᵢ|| ∀ i

KuoLeeUtils.rankfact — Method

Rankfact

Factorize a matrix in rank one matrices via the singular value decomposition.

\[ T = EF^T, \quad E \in \mathbb{F}^{n_2n_3 \times r}, \quad F\in \mathbb{F}^{n_1\times R} \]

\[ T = U \Sigma V^T, \quad E = U, \quad F^T = \Sigma V^T.\]

Inputs

T: two-dimensonal array to be factorized

r: number of rank one terms

Output

E: E and F are the factor matrices of the factorization

F: E anf F are the factor matrices of the factorization

Utils.condtdp — Method

condtdp

Calculate the condition number of the tensor decomposition problem based on Terracini's matrix.

Inputs

X

Y

Z

dims

R

Utils.cpd_gen — Method

cpd_gen(dims, r)

Generate a flattened tensor with dimensions n₂n₃×n₁ of rank r. The tensor is generated from random factor matrices as

\[ T = (Z \odot Y)X^H. \]

Inputs

dims: tuple with the dimensions (n₁, n₂, n₃).

r: the desired tensor rank

Output

X: factor matrix of dimension n₁×r

Y: factor matrix of dimension n₂×r

Z: factor matrix of dimension n₃×r

T: flattened tensor n₂n₃×n₁

Utils.cpd_gen — Method

cpd_gen(X, Y, Z)

Generate a flattened tensor from the given factor matrices such that

\[ T = (Z \odot Y)X^H. \]

Utils.signnorm — Method

signnorm

Rescale matrix columns such that the first row has positive sign. In the case of complex entries, the elements on the first row have zero complex angle.

Examples

>> A
>> 2×2 Array{Complex{Float64},2}:
    1.30736-1.34255im    0.183123-0.504201im
    -0.00153446+0.0305201im  0.222609-0.0412371im
>> signnorm(A)
>> 2×2 Array{Complex{Float64},2}:
    1.87393+1.5908e-16im  0.536426-2.60882e-17im
    -0.0229362+0.0201932im   0.114753+0.195159im